+++ docs/csharp/language-reference/statements/snippets/jump-statements/RefParameterModifier.cs
@@ -6,27 +6,6 @@ namespace InRefOutModifier
 {
     public class RefParameterModifier
     {
-        public static void Examples()
-        {
-            FirstRefExample();
-            ModifyProductsByReference();
-        }
-
-        private static void FirstRefExample()
-        {
-            // <Snippet1>
-            void Method(ref int refArgument)
-            {
-                refArgument = refArgument + 44;
-            }
-
-            int number = 1;
-            Method(ref number);
-            Console.WriteLine(number);
-            // Output: 45
-            // </Snippet1>
-        }
-
         // <SnippetFindReturningRef>
         public static ref int Find(int[,] matrix, Func<int, bool> predicate)
         {
@@ -38,48 +17,6 @@ namespace InRefOutModifier
         }
         // </SnippetFindReturningRef>
 
-
-        //<Snippet3>
-        class Product
-        {
-            public Product(string name, int newID)
-            {
-                ItemName = name;
-                ItemID = newID;
-            }
-
-            public string ItemName { get; set; }
-            public int ItemID { get; set; }
-        }
-
-        private static void ChangeByReference(ref Product itemRef)
-        {
-            // Change the address that is stored in the itemRef parameter.
-            itemRef = new Product("Stapler", 99999);
-
-            // You can change the value of one of the properties of
-            // itemRef. The change happens to item in Main as well.
-            itemRef.ItemID = 12345;
-        }
-
-        private static void ModifyProductsByReference()
-        {
-            // Declare an instance of Product and display its initial values.
-            Product item = new Product("Fasteners", 54321);
-            System.Console.WriteLine("Original values in Main.  Name: {0}, ID: {1}\n",
-                item.ItemName, item.ItemID);
-
-            // Pass the product instance to ChangeByReference.
-            ChangeByReference(ref item);
-            System.Console.WriteLine("Back in Main.  Name: {0}, ID: {1}\n",
-                item.ItemName, item.ItemID);
-        }
-
-        // This method displays the following output:
-        // Original values in Main.  Name: Fasteners, ID: 54321
-        // Back in Main.  Name: Stapler, ID: 12345
-
-        // </Snippet3>
 
         private static void BookCollectionExample()
         {
@@ -100,14 +37,7 @@ namespace InRefOutModifier
             // </Snippet5>
         }
     }
-    //<Snippet2>
-    class RefOverloadExample
-    {
-        public void SampleMethod(int i) { }
-        public void SampleMethod(ref int i) { }
-    }
-    // </Snippet2>
-
+    
     // <Snippet4>
 
     public class Book
